requirements.txt
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
bidict==0.23.1
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.7
dnspython==2.7.0
eventlet==0.39.1
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
psycopg==3.2.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
redis==5.2.1
service-identity==24.2.0
setuptools==78.1.0
simple-websocket==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
zope.interface==7.2
